Use SHA for pulling docker reference instead of `latest`
This ensures that we can still use the "latest" image, but have a clear
reference of the container we're actually building from. This helps us
not have to rely on a full release workflow for the base image since it
just pulls from krakend and won't be updated as much.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>